aws-jupyter/requirements.txt
awscli==1.15.39
PyYAML==3.10
